Modified: 更改指针类型，主要是ios端指针是更大的数据类型。为手动调用缩放模型添加刷新。

+++ OSG/inc/public/PointerAccess.h
@@ -22,7 +22,7 @@ BOOL IsBadWritePtr(const void *lp, size_t ucb);
 
 void /*__cdecl*/ PtrTestHandler(int nSig);
 
-BOOL IsBadReadPtr(void* lp, UINT cb);
+BOOL IsBadReadPtr(void* lp, size_t cb);
 
 #endif
